fix: Replace SignUpForm with SignUpFormWithButton and streamline sign-up process to  fix triggering checkbox state
