+++ lib/features/auth/sign_up/ui/widgets/sign_up_body.dart
@@ -1,14 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:nexus/core/helpers/base_extensions/context/navigation.dart';
 import 'package:nexus/core/widgets/have_account_question_text.dart';
 import 'package:nexus/core/helpers/helper_methods/spacing.dart';
-import 'package:nexus/features/auth/sign_up/ui/widgets/sign_up_form.dart';
+import 'package:nexus/features/auth/sign_up/ui/widgets/sign_up_form_with_button.dart';
 import 'package:nexus/features/auth/sign_up/ui/widgets/sign_up_header.dart';
 
 import '../../../../../core/routing/routes.dart';
-import '../../../../../core/widgets/app_text_button.dart';
-import '../../logic/sign_up_cubit.dart';
 import 'sign_up_bloc_listener.dart';
 
 class SignUpBody extends StatelessWidget {
@@ -25,14 +22,7 @@ class SignUpBody extends StatelessWidget {
           verticalSpace(24),
           const SignUpHeader(),
           verticalSpace(48),
-          const SignUpForm(),
-          verticalSpace(16),
-          AppTextButton(
-            onPressed: () {
-              validateThenSignUp(context);
-            },
-            text: 'Sign Up',
-          ),
+          const SignUpFormWithButton(),
           verticalSpace(24),
           HaveAccountQuestionText(
             questionText: 'Already have an account? ',
@@ -46,15 +36,5 @@ class SignUpBody extends StatelessWidget {
         ],
       ),
     );
-  }
-
-  void validateThenSignUp(BuildContext context) {
-    context.read<SignUpCubit>().isTermsValid =
-        context.read<SignUpCubit>().isTermsChecked;
-
-    if (context.read<SignUpCubit>().formkey.currentState!.validate() &&
-        context.read<SignUpCubit>().isTermsValid) {
-      context.read<SignUpCubit>().emitSignUpStates();
-    }
   }
 }
